remove and edit endpoints

+++ backend/auth/auth.go
@@ -10,7 +10,7 @@ import (
 )
 
 type User struct {
-	ID        int64  `json:"id"`
+	ID        int    `json:"id"`
 	Username  string `json:"user"`
 	password  string
 	apikey    *string
@@ -137,4 +137,3 @@ func newUUID() (string, error) {
 	uuid[6] = uuid[6]&^0xf0 | 0x40
 	return fmt.Sprintf("%x-%x-%x-%x-%x", uuid[0:4], uuid[4:6], uuid[6:8], uuid[8:10], uuid[10:]), nil
 }
-
